+++ malachite-nz/src/gaussian_integer/conversion/string/mod.rs
@@ -9,5 +9,11 @@
 /// Implementations of traits for converting strings to
 /// [`GaussianInteger`](crate::gaussian_integer::GaussianInteger)s.
 pub mod from_string;
+/// The implementation of [`ToLatex`](malachite_base::strings::latex::ToLatex) for a
+/// [`GaussianInteger`].
+pub mod latex;
 /// Implementations of traits for converting to strings.
 pub mod to_string;
+/// The implementation of [`ToTypst`](malachite_base::strings::typst::ToTypst) for a
+/// [`GaussianInteger`].
+pub mod typst;